+++ app/Filament/Pages/Auth/Login.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Filament\Standard\Pages\Auth;
+namespace App\Filament\Pages\Auth;
 
 use Filament\Auth\Pages\Login as BaseLogin;
 use Filament\Schemas\Components\Component;
@@ -11,9 +11,9 @@ use Illuminate\Contracts\Support\Htmlable;
 
 class Login extends BaseLogin
 {
-    protected string $view = 'filament.standard.pages.auth.login';
+    protected string $view = 'filament.pages.auth.login';
 
-    protected static string $layout = 'filament.standard.layouts.login';
+    protected static string $layout = 'filament.layouts.login';
 
     public function getHeading(): string|Htmlable|null
     {